+++ ui/app/src/modules/System/Publishing/Queue/PluginDetailsView.tsx
