+++ ui/app/src/modules/System/Sites/Create/BlueprintCard.tsx
