+++ ui/app/src/utils/codebase-bridge.tsx
@@ -101,8 +101,8 @@ export function createCodebaseBridge() {
             default: module.DependencySelectionDelete
           }))
       )),
-      CreateSiteDialog: lazy(() => import('../modules/System/Sites/create/CreateSiteDialog')),
-      PublishingQueue: lazy(() => import('../modules/System/Publishing/queue/PublishingQueue'))
+      CreateSiteDialog: lazy(() => import('../modules/System/Sites/Create/CreateSiteDialog')),
+      PublishingQueue: lazy(() => import('../modules/System/Publishing/Queue/PublishingQueue'))
     },
 
     assets: {
